requirements.txt
tensorflow==1.13.1
keras==2.2.4
tqdm==4.54.1
scikit-learn==0.23.2
tensorboard==1.13.1
opencv==3.4.2
opencv-python==4.5.5.64
minio==7.1.0
h5py==2.10.0
pandas==1.1.5
numpy==1.16.0
pillow==8.2.0
pymysql==0.10.1
redis==3.5.3
matplotlib==3.3.4
pydot==1.4.1
scikit-image==0.17.2
pyopenssl==22.0.0
dnspython==2.2.1